New design for Residuals & BinaryConfusionMatrix

+++ spark_adatao/src/test/java/com/adatao/spark/ddf/analytics/MetricsTests.java
@@ -39,13 +39,11 @@ public class MetricsTests {
           "distance, depdelay, if (arrdelay > 10.89, 1, 0) as delayed from airline");
     Assert.assertEquals(3, ddf.getSummary().length);
     IModel logModel = ddf.ML.train("logisticRegressionWithSGD", 10, 0.1);
-    Long[][] cm = ddf.ML.getConfusionMatrix(logModel, 0.5);
-    Assert.assertEquals(0, cm[0][0].longValue());
-    Assert.assertEquals(13, cm[0][1].longValue());
-    Assert.assertEquals(0, cm[1][0].longValue());
-    Assert.assertEquals(18, cm[1][1].longValue());
+    long[][] cm = ddf.ML.getConfusionMatrix(logModel, 0.5);
+    Assert.assertEquals(0, cm[0][0]);
+    Assert.assertEquals(13, cm[0][1]);
+    Assert.assertEquals(0, cm[1][0]);
+    Assert.assertEquals(18, cm[1][1]);
     manager.shutdown();
   }
-
-
 }
